+++ puppeteerServer/util.ts
@@ -10,7 +10,7 @@ export function match(string:string, subStrings:string[]) {
 
 export function validMangaCheck(url:string):mangaUrlCheck {
     if (!url.includes('http')) return {success: false, value: 'invalid URL Format!', statusCode: 422}
-    if (!url.includes('chapter') && !url.includes('ch-')) return {success: false, value: 'link provided is for an overview page. Please provide a link to a specific chapter page!', statusCode: 422}
+    if (!checkChapterUrl(url)) return {success: false, value: 'link provided is for an overview page. Please provide a link to a specific chapter page!', statusCode: 422}
     
     if (url.includes('manganato') && config.sites.allowManganatoScans) return {success: true, value: 'manganato'}
     else if (url.includes('mangadex') && config.sites.allowMangaDex) return {success: true, value: 'mangadex'}
@@ -19,4 +19,13 @@ export function validMangaCheck(url:string):mangaUrlCheck {
     else if (url.includes('asura') && config.sites.allowAsura) return {success: true, value: 'asura'}
     else if (url.includes('comick') && config.sites.allowComick) return {success: true, value: 'comick'}
     else return {success: false, value: 'Unsupported WebPage!', statusCode: 422}
+}
+
+function checkChapterUrl (url:string):boolean {
+    url = url.toLowerCase()
+    if (url.includes('comick.io/comic/')) {
+        url.replace('https://comick.io/comic/','')
+        if (url.includes('/')) return true
+    }
+    return (url.includes('chapter') || url.includes('ch-'))
 }